refactor types + transforms
